requirements.txt
emoji==2.0.0
loguru==0.6.0
Pillow==9.2.0
pilmoji==2.0.2
PyYAML==5.1